requirements.txt
telebot
telegram
python-telegram-bot==11.1.0
python-dotenv==0.10.3
vk-api==11.4.0
redis==3.3.8

